+++ cognite/experimental/data_classes/functions.py
@@ -1,4 +1,5 @@
-from typing import Dict, List, Union
+import time
+from typing import Dict, List, Optional, Union
 
 from cognite.client.data_classes._base import CogniteResource, CogniteResourceList
 
@@ -49,11 +50,19 @@ class Function(CogniteResource):
         self.error = error
         self._cognite_client = cognite_client
 
-    def call(self, data=None, asynchronous: bool = False):
-        return self._cognite_client.functions.call(id=self.id, data=data, asynchronous=asynchronous)
+    def call(self, data=None, wait: bool = True):
+        return self._cognite_client.functions.call(id=self.id, data=data, wait=wait)
 
-    def list_calls(self):
-        return self._cognite_client.functions.calls.list(function_id=self.id)
+    def list_calls(
+        self,
+        status: Optional[str] = None,
+        schedule_id: Optional[int] = None,
+        start_time: Optional[Dict[str, int]] = None,
+        end_time: Optional[Dict[str, int]] = None,
+    ):
+        return self._cognite_client.functions.calls.list(
+            function_id=self.id, status=status, schedule_id=schedule_id, start_time=start_time, end_time=end_time
+        )
 
     def list_schedules(self):
         all_schedules = self._cognite_client.functions.schedules.list()
@@ -116,8 +125,8 @@ class FunctionCall(CogniteResource):
         id (int): A server-generated ID for the object.
         start_time (int): The number of milliseconds since 00:00:00 Thursday, 1 January 1970, Coordinated Universal Time (UTC), minus leap seconds.
         end_time (int): The number of milliseconds since 00:00:00 Thursday, 1 January 1970, Coordinated Universal Time (UTC), minus leap seconds.
-        response (str): Response from the function. The function must return a JSON serializable object.
         status (str): Status of the function call ("Running" or "Completed").
+        schedule_id (int): The schedule id belonging to the call.
         error (dict): Error from the function call. It contains an error message and the stack trace.
         cognite_client (CogniteClient): An optional CogniteClient to associate with this data class.
     """
@@ -127,8 +136,8 @@ class FunctionCall(CogniteResource):
         id: int = None,
         start_time: int = None,
         end_time: int = None,
-        response: str = None,
         status: str = None,
+        schedule_id: int = None,
         error: dict = None,
         function_id: int = None,
         cognite_client=None,
@@ -136,33 +145,50 @@ class FunctionCall(CogniteResource):
         self.id = id
         self.start_time = start_time
         self.end_time = end_time
-        self.response = response
         self.status = status
+        self.schedule_id = schedule_id
         self.error = error
-        self._function_id = function_id
+        self.function_id = function_id
         self._cognite_client = cognite_client
 
-    def get_logs(self):
-        return self._cognite_client.functions.calls.get_logs(call_id=self.id, function_id=self._function_id)
+    def get_response(self):
+        return self._cognite_client.functions.calls.get_response(call_id=self.id, function_id=self.function_id)
 
-    @classmethod
-    def _load(cls, resource: Union[Dict, str], function_id: int = None, cognite_client=None):
-        instance = super()._load(resource, cognite_client=cognite_client)
-        if function_id:
-            instance._function_id = function_id
-        return instance
+    def get_logs(self):
+        return self._cognite_client.functions.calls.get_logs(call_id=self.id, function_id=self.function_id)
+
+    def update(self):
+        latest = self._cognite_client.functions.calls.retrieve(call_id=self.id, function_id=self.function_id)
+        self.status = latest.status
+        self.end_time = latest.end_time
+        self.error = latest.error
+
+    def wait(self):
+        while self.status == "Running":
+            self.update()
+            time.sleep(1.0)
 
 
 class FunctionCallList(CogniteResourceList):
     _RESOURCE = FunctionCall
     _ASSERT_CLASSES = False
 
-    @classmethod
-    def _load(cls, resource: Union[List, str], function_id: int, cognite_client=None):
-        instance = super()._load(resource, cognite_client=cognite_client)
-        for obj in instance:
-            obj._function_id = function_id
-        return instance
+
+class FunctionCallResponse(CogniteResource):
+    """The response from a function call.
+
+    Args:
+        function_id (int): The ID of the function on which the call was made.
+        call_id (int): The ID of the call.
+        response (str): The function call response.
+        cognite_client (CogniteClient): An optional CogniteClient to associate with this data class.
+    """
+
+    def __init__(self, call_id: int = None, function_id: int = None, response: str = None, cognite_client=None):
+        self.call_id = call_id
+        self.function_id = function_id
+        self.response = response
+        self._cognite_client = cognite_client
 
 
 class FunctionCallLogEntry(CogniteResource):